requirements.txt
numpy
PyYAML
scikit-learn
streamlit
langchain-community
langchain-core
langchain-text-splitters
rank-bm25
requests
beautifulsoup4
lxml
